feat: add UserProfileRepositoryInterface and inject it in UserProfileService
